- removed compile warning (float vs. double & int vs. float)
git-svn-id: svn://mirror.orbiter-radio.co.uk/D3D9client/trunk@187 c0f5cd98-1510-604a-8459-bf7538b8c9c2

+++ Orbitersdk/D3D9Client/DebugControls.cpp
@@ -429,7 +429,7 @@ void SetColorSlider()
 	float val = GetMaterialValue(MatPrp, SelColor);
 	
 	//if (MatPrp==2 && SelColor==3) val/=1000.0; // Specular Power
-	if (MatPrp==2 && SelColor==3) val = log(val) / 6.907755279; // Specular Power
+	if (MatPrp==2 && SelColor==3) val = float(log(val) / 6.907755279); // Specular Power
 	if (MatPrp==5 && SelColor==0) val/=12.0;  // Dissolve scale
 	if (MatPrp==5 && SelColor==1) val/=0.2f;  // Dissolve scatter
 	if (MatPrp==6 && SelColor==0) val/=6.0f;  // Fresnel range
@@ -565,7 +565,7 @@ void UpdateColorSlider(WORD pos)
 	if (SelColor==3) bLink = false;
 
 	//if (MatPrp==2 && SelColor==3) val*=1000.0; // Specular Power
-	if (MatPrp==2 && SelColor==3) val = exp(val*6.907755279); // Specular Power
+	if (MatPrp==2 && SelColor==3) val = float(exp(val*6.907755279)); // Specular Power
 	if (MatPrp==5 && SelColor==0) val*=12.0;  // Dissolve scale
 	if (MatPrp==5 && SelColor==1) val*=0.2f;  // Dissolve scatter
 	if (MatPrp==6 && SelColor==0) val*=6.0f;  // Fresnel range
